Financial_BudgetAnalysis
Each country has a financial budget that describes the government’s spending capacity in different sectors of the economy.

There are so many Data Analysts today that come from a non-coding background. If you are from a commerce background then you may know what is a financial budget. In short, it is a detailed report on the income and expenditure of the government for a financial year.

You may get the task of analyzing a country’s financial budget every year if you are working as a data analyst in the media and communications field, as the media have to explain the government’s priorities for the complete financial year.