Update README.md
Conclusion
We can see that the finance department is getting 40% of the funds. So this is how we can analyze a dataset that contains data about the revenue and expenditure of the government for a financial year. I hope you liked this article on Financial Budget analysis with Python. Feel free to ask your valuable questions in the comments section below.